[CSL-2346] Pos.Ntp.Configuration moved from Pos.Infra
Also move `Pos.NtpCheck` to `Pos.Ntp.Check`.
