+++ infra/Pos/Ntp.hs
@@ -1,0 +1,3 @@
+-- | Reexports of error reporting functionality
+
+{-# OPTIONS_GHC -F -pgmF autoexporter #-}
