+++ lib/src/Test/Pos/Configuration.hs
@@ -28,7 +28,7 @@ import           Pos.Core (BlockVersionData, HasConfiguration, withGenesisSpec)
 import           Pos.Core.Configuration (CoreConfiguration (..), GenesisConfiguration (..))
 import           Pos.Core.Genesis (GenesisSpec (..))
 import           Pos.Delegation (HasDlgConfiguration, withDlgConfiguration)
-import           Pos.Infra.Configuration (HasNtpConfiguration, withNtpConfiguration)
+import           Pos.Ntp.Configuration (HasNtpConfiguration, withNtpConfiguration)
 import           Pos.Launcher.Configuration (Configuration (..), HasConfigurations)
 import           Pos.Ssc.Configuration (HasSscConfiguration, withSscConfiguration)
 import           Pos.Txp (HasTxpConfiguration, withTxpConfiguration)
