+++ wallet/src/Pos/Wallet/Web/Methods/Misc.hs
@@ -51,8 +51,8 @@ import           Pos.Client.KeyStorage (MonadKeys (..), deleteAllSecretKeys)
 import           Pos.Configuration (HasNodeConfiguration)
 import           Pos.Core (Address, HasConfiguration, SlotId, SoftwareVersion (..))
 import           Pos.Crypto (hashHexF)
-import           Pos.Infra.Configuration (HasNtpConfiguration, ntpConfiguration)
-import           Pos.NtpCheck (NtpCheckMonad, NtpStatus (..), getNtpStatusOnce)
+import           Pos.Ntp.Configuration (HasNtpConfiguration, ntpConfiguration)
+import           Pos.Ntp.Check (NtpCheckMonad, NtpStatus (..), getNtpStatusOnce)
 import           Pos.Shutdown (HasShutdownContext, triggerShutdown)
 import           Pos.Slotting (MonadSlots, getCurrentSlotBlocking)
 import           Pos.Txp (TxId, TxIn, TxOut)
